move actual function running into hook.c
